requirements.txt
streamlit
fastapi
uvicorn
python-dotenv
google-generativeai
groq
pydantic
requests

